remove type checking from insert analyzer
