+++ Cargo.toml
@@ -17,16 +17,17 @@ members = [
     "deprecated/query_planning_deprecated/query_planner",
     "deprecated/schema_executor",
     "deprecated/sql_model",
+    "entities/types",
+    "entities/definition",
     "meta_def",
     "query_analysis/expr_operators",
     "query_analysis/query_analyzer",
     "query_analysis/tree",
     "query_parsing/parser",
     "query_parsing/sql-ast",
-    "repr",
+    "entities/repr",
     "server/connection",
     "server/pg_model",
     "server/node",
-    "types",
     "write_query/executor",
 ]
