+++ data/binary/Cargo.toml
@@ -6,4 +6,4 @@ edition = "2018"
 publish = false
 
 [dependencies]
-repr = { path = "../../repr" }
+repr = { path = "../../entities/repr" }
