+++ data/definition_operations/Cargo.toml
@@ -6,4 +6,4 @@ edition = "2018"
 publish = false
 
 [dependencies]
-types = { path = "../../types" }
+types = { path = "../../entities/types" }
