+++ deprecated/ast/Cargo.toml
@@ -7,8 +7,8 @@ publish = false
 
 [dependencies]
 meta_def = { path = "../../meta_def" }
-repr = { path = "../../repr" }
-types = { path = "../../types" }
+repr = { path = "../../entities/repr" }
+types = { path = "../../entities/types" }
 sql_ast = { path = "../../query_parsing/sql-ast" }
 
 bigdecimal = { version = "0.2.0", features = ["string-only"] }
