+++ deprecated/query_analysis_deprecated/description/Cargo.toml
@@ -7,7 +7,7 @@ publish = false
 
 [dependencies]
 meta_def = { path = "../../../meta_def" }
-types = { path = "../../../types" }
+types = { path = "../../../entities/types" }
 pg_wire = "0.5.0"
 
 sql_ast = { path = "../../../query_parsing/sql-ast" }
