+++ deprecated/query_execution_deprecated/expr_eval/Cargo.toml
@@ -7,7 +7,7 @@ publish = false
 
 [dependencies]
 ast = { path = "../../ast" }
-repr = { path = "../../../repr" }
+repr = { path = "../../../entities/repr" }
 
 bigdecimal = { version = "0.2.0", features = ["string-only"] }
