+++ deprecated/query_planning_deprecated/plan/Cargo.toml
@@ -11,6 +11,6 @@ constraints = { path = "../../constraints_deprecated" }
 meta_def = { path = "../../../meta_def" }
 pg_wire = "0.5.0"
 sql_model = { path = "../../sql_model" }
-types = { path = "../../../types" }
+types = { path = "../../../entities/types" }
 
 sql_ast = { path = "../../../query_parsing/sql-ast" }
