+++ entities/definition/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "definition"
+version = "0.1.0"
+authors = ["Alex Dukhno <alex.dukhno@icloud.com>"]
+edition = "2018"
+publish = false
+
+[dependencies]
+repr = { path = "../repr" }
+types = { path = "../types" }
+sql_ast = { path = "../../query_parsing/sql-ast" }
