+++ query_analysis/query_analyzer/Cargo.toml
@@ -10,7 +10,8 @@ analysis_tree = { path = "../tree" }
 catalog = { path = "../../data/catalog" }
 expr_operators = { path = "../expr_operators" }
 meta_def = { path = "../../meta_def" }
-types = { path = "../../types" }
+definition = { path = "../../entities/definition" }
+types = { path = "../../entities/types" }
 sql_ast = { path = "../../query_parsing/sql-ast" }
 
 # deprecated module
