+++ query_analysis/tree/Cargo.toml
@@ -8,5 +8,6 @@ publish = false
 [dependencies]
 expr_operators = { path = "../expr_operators" }
 meta_def = { path = "../../meta_def" }
-types = { path = "../../types" }
+types = { path = "../../entities/types" }
+definition = { path = "../../entities/definition" }
 sql_ast = { path = "../../query_parsing/sql-ast" }
